+++ Python/optimizer_cases.c.h
@@ -980,7 +980,7 @@
                 ctx->done = true;
                 break;
             }
-            int returning_stacklevel = this_instr->operand1;
+            int returning_stacklevel = (int)this_instr->operand1;
             if (ctx->curr_frame_depth >= 2) {
                 PyCodeObject *expected_code = ctx->frames[ctx->curr_frame_depth - 2].code;
                 if (expected_code == returning_code) {
@@ -1049,7 +1049,7 @@
                 break;
             }
             _Py_BloomFilter_Add(dependencies, returning_code);
-            int returning_stacklevel = this_instr->operand1;
+            int returning_stacklevel = (int)this_instr->operand1;
             if (frame_pop(ctx, returning_code, returning_stacklevel)) {
                 break;
             }
@@ -3015,7 +3015,7 @@
                 break;
             }
             _Py_BloomFilter_Add(dependencies, returning_code);
-            int returning_stacklevel = this_instr->operand1;
+            int returning_stacklevel = (int)this_instr->operand1;
             if (frame_pop(ctx, returning_code, returning_stacklevel)) {
                 break;
             }
